+++ src/datatypes/account.ts
@@ -3,4 +3,5 @@ import type { components } from "@/schema/api.js";
 type Schemas = components["schemas"];
 
 export type Account = Schemas["Account"];
+export type PublicAccount = Schemas["PublicAccount"];
 export type UpdateAccount = Schemas["UpdateAccount"];